# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Arletta Gym API",
        "description": "REST API for Arletta Gym Landing Page — Authentication, Memberships, Trainers, Schedules, Bookings, Payments, Signatures",
        "contact": {
            "name": "Arletta Gym Developer",
            "email": "admin@arlettagym.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Local Development"
        }
    ],
    "paths": {
...
        "/api/trainers": {
            "get": {
                "tags": [
                    "Trainers"
                ],
                "summary": "List all trainers",
                "operationId": "9ebe45c9993bb130b430d3409565c758",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "gym_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trainers retrieved"
                    }
                }
            }
        },
        "/api/trainers/{id}": {
            "get": {
                "tags": [
                    "Trainers"
                ],
                "summary": "Get trainer detail with packages",
                "operationId": "8eee0bbcecdcee80c983b6af3c08fb8f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trainer detail retrieved"
                    }
                }
            }
        },
        "/api/trainers/{id}/clients": {
            "get": {
                "tags": [
                    "Trainers"
                ],
                "summary": "Get trainer's clients (PT only)",
                "operationId": "dda25486ace83bb108f2456be63cb93a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Clients retrieved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/trainers/clients/all": {
            "get": {
                "tags": [
                    "Trainers"
                ],
                "summary": "Get all gym members (PT views all, not just their clients)",
                "operationId": "db3bafe4c486a7d17ea400a830b2ea10",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All members retrieved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/trainers/{id}/schedules": {
            "get": {
                "tags": [
                    "Trainers"
                ],
                "summary": "Get trainer's class schedules",
                "operationId": "c7e4435f741011e3ed864511e365b1f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trainer schedules retrieved"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}